requirements.txt
aiogram
asyncio
python-telegram-bot-calendar